+++ turbo/packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [8.559.1](https://github.com/vm0-ai/vm0/compare/core-v8.559.0...core-v8.559.1) (2026-08-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.329.7
+
 ## [8.559.0](https://github.com/vm0-ai/vm0/compare/core-v8.558.0...core-v8.559.0) (2026-08-13)